requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
anyio==3.7.1
astunparse==1.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==45.0.7
fastapi==0.103.2
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jax==0.4.30
jaxlib==0.4.30
joblib==1.5.3
keras==3.12.1
Keras-Preprocessing==1.1.2
numpy==2.2.6
opencv-python-headless==4.11.0.86
opt-einsum==3.4.0
pandas==2.3.3
pillow==12.1.0
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.15.3
tensorboard==2.20.0
tensorflow==2.20.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.5
